Altered Building capacity
Building capacity is now an array of capacities, corresponding to the array of uses, and then stored in an array with fixed length of Use.values().length. Meaning when an instance of the BuildingName enum is created, e.g. piazza, the uses can be listed in any order, and uses[0] will be given the capacity in capacity[0]. But when it is stored in this.capacity, there is always the same number of values in this.capacity, as there are uses in Use, so it's stored as this.capacity[use] = capacity[0] for example. Meaning accessing it later in the code is always the same order.