Changes and fixes to satisfaction
Made score caps private
Changed how updateAverageDistances checks for comparison against a the same building so that it now works properly
Added a check for when removing a building has left only 1 of that type on the map, or 0 on the map, this now sets the average distance and count to 0 when that occurs. This prevents the div 0 which was causing distances to be set to infinity, - infinity or infinity - infinity, which gave NaN,
Changed how updateBuildingDistancesScore interacts with calculateScoreBonus, the check preventing adding score when for buildings that weren't placed, or for the distance between buildings of the same type when only one of that type was placed, is now handled in calculateScoreBonus. This fixed an issue where score would increase, but never decrease. Basically it was easier to always add a score, but check when the score should be 0, than check when not to add a score.
Altered values in the weight matrix such that it actually adds up to 40 now (not 56).